Add logic to calendar
