+++ src/components/Reservation/Modal.js
@@ -1,8 +1,15 @@
 /* eslint-disable react/prop-types */
-import React, { useRef, useEffect, useCallback } from 'react';
+import React, {
+  useRef, useEffect, useCallback, useState,
+} from 'react';
+import { useLocation } from 'react-router-dom';
 import { useSpring, animated } from 'react-spring';
+import { useSelector, useDispatch } from 'react-redux';
 import styled from 'styled-components';
 import { MdClose } from 'react-icons/md';
+import DatePicker from 'react-datepicker';
+import { getItems } from '../../api/items';
+import reservedItems from '../../api/reservedItems';
 
 const Background = styled.div`
   width: 100%;
@@ -20,19 +27,19 @@ const ModalWrapper = styled.div`
   box-shadow: 0 5px 16px rgba(0, 0, 0, 0.2);
   background: #fff;
   color: #000;
-  display: grid;
-  grid-template-columns: 1fr 1fr;
+  display: flex;
+  flex-direction: column;
   position: relative;
   z-index: 10;
   border-radius: 10px;
 `;
 
-const ModalImg = styled.img`
-  width: 100%;
-  height: 100%;
-  border-radius: 10px 0 0 10px;
-  background: #000;
-`;
+// const ModalImg = styled.img`
+//   width: 100%;
+//   height: 100%;
+//   border-radius: 10px 0 0 10px;
+//   background: #000;
+// `;
 
 const ModalContent = styled.div`
   display: flex;
@@ -41,6 +48,7 @@ const ModalContent = styled.div`
   align-items: center;
   line-height: 1.8;
   color: #141414;
+  margin-top: 7%;
   p {
     margin-bottom: 1rem;
   }
@@ -61,11 +69,22 @@ const CloseModalButton = styled(MdClose)`
   height: 32px;
   padding: 0;
   z-index: 10;
+  margin-right: 5%;
 `;
 
 const Modal = ({ showModal, setShowModal }) => {
+  const items = useSelector((state) => state.items.items) || [];
+  const [loginResponse, setLoginResponse] = useState('');
   const modalRef = useRef();
+  const dispatch = useDispatch();
+  const location = useLocation();
 
+  const token = JSON.parse(localStorage.getItem('token'));
+  useEffect(() => {
+    dispatch(getItems(token));
+  }, []);
+  const [startDate, setStartDate] = useState(new Date());
+  const [endDate, setEndDate] = useState(new Date());
   const animation = useSpring({
     config: {
       duration: 250,
@@ -77,6 +96,36 @@ const Modal = ({ showModal, setShowModal }) => {
   const closeModal = (e) => {
     if (modalRef.current === e.target) {
       setShowModal(false);
+    }
+  };
+  const getItemId = () => {
+    if (location.state) {
+      return location.state.id;
+    }
+    if (items[0]) {
+      return items[0].id;
+    }
+    return 1;
+  };
+  const [itemId, setItemId] = useState(getItemId());
+  const user = useSelector((state) => state.usersReducer.user);
+
+  const reserveSubmit = async (e) => {
+    e.preventDefault();
+    if (endDate < new Date() || startDate === new Date()) {
+      setLoginResponse('Please enter a correct date');
+    } else {
+      const response = await reservedItems({
+        start_date: startDate,
+        end_date: endDate,
+        user_id: user.id,
+        id: parseInt(itemId, 10),
+      });
+      if (!response.error) {
+        setLoginResponse('Succesfully Reserved');
+      } else {
+        setLoginResponse(response.error);
+      }
     }
   };
 
@@ -104,13 +153,40 @@ const Modal = ({ showModal, setShowModal }) => {
         <Background onClick={closeModal} ref={modalRef}>
           <animated.div style={animation}>
             <ModalWrapper showModal={showModal}>
-              <ModalImg src="modal.jpg" alt="camera" />
               <ModalContent>
-                <h1>Are you ready?</h1>
-                <p>Get exclusive access to our next launch.</p>
-                <div className="d-flex w-100 justify-content-center pt-5">
-                  <button className="btn btn-outline-success rounded-pill" type="submit">Reserve</button>
-                </div>
+                <form className="d-flex flex-column h-100 justify-content-center align-items-center " onSubmit={reserveSubmit}>
+                  <section>
+                    <span>
+                      {loginResponse}
+                    </span>
+                    <div className="d-flex flex-column justify-content-center text-white">
+                      <h2 className="text-center">Reserve an Arena</h2>
+                    </div>
+                    <div className="d-flex">
+                      <select className="form-select me-2 rounded-pill" onChange={(e) => setItemId(e.target.value)} value={itemId}>
+                        {items.map((item) => (
+
+                          <option key={item.id} value={item.id}>
+                            {item.name}
+
+                          </option>
+
+                        ))}
+                      </select>
+
+                    </div>
+                    <div className="d-flex justify-content-center w-100 mt-3">
+                      <DatePicker selected={startDate} onChange={(date) => setStartDate(date)} />
+                    </div>
+                    <div className="d-flex justify-content-center w-100 mt-3">
+                      <DatePicker selected={endDate} onChange={(date) => setEndDate(date)} />
+                    </div>
+                    <div className="d-flex w-100 justify-content-center pt-5">
+                      <button className="btn btn-outline-success rounded-pill" type="submit">Reserve</button>
+                    </div>
+                  </section>
+                </form>
+
               </ModalContent>
               <CloseModalButton
                 aria-label="Close modal"
