+++ src/components/Reservation/ReservationForm.js
@@ -1,11 +1,8 @@
 import React, { useState, useEffect } from 'react';
-import { useSelector, useDispatch } from 'react-redux';
-import { useLocation } from 'react-router-dom';
-import DatePicker from 'react-datepicker';
+import { useDispatch } from 'react-redux';
 import styled from 'styled-components';
 import { getItems } from '../../api/items';
 import Modal from './Modal';
-import reservedItems from '../../api/reservedItems';
 import 'react-datepicker/dist/react-datepicker.css';
 import GlobalStyle from './globalStyles';
 import './Reserve.scss';
@@ -28,11 +25,6 @@ const Button = styled.button`
 `;
 
 const ReservationForm = () => {
-  const items = useSelector((state) => state.items.items) || [];
-  const location = useLocation();
-  const [startDate, setStartDate] = useState(new Date());
-  const [endDate, setEndDate] = useState(new Date());
-  const [loginResponse, setLoginResponse] = useState('');
   const [showModal, setShowModal] = useState(false);
   const openModal = () => {
     setShowModal((prev) => !prev);
@@ -44,67 +36,14 @@ const ReservationForm = () => {
     dispatch(getItems(token));
   }, []);
 
-  const getItemId = () => {
-    if (location.state) {
-      return location.state.id;
-    }
-    if (items[0]) {
-      return items[0].id;
-    }
-    return 1;
-  };
-
-  const [itemId, setItemId] = useState(getItemId());
-  const user = useSelector((state) => state.usersReducer.user);
-
-  const reserveSubmit = async (e) => {
-    e.preventDefault();
-    const response = await reservedItems({
-      start_date: startDate,
-      end_date: endDate,
-      user_id: user.id,
-      id: parseInt(itemId, 10),
-    });
-    if (!response.error) {
-      setLoginResponse('Succesfully Reserved');
-    } else {
-      setLoginResponse(response.error);
-    }
-  };
-
   return (
     <div className="wrapper">
-      <form className="d-flex flex-column h-100 justify-content-center align-items-center " onSubmit={reserveSubmit}>
-        <section>
-          <span>
-            {loginResponse}
-          </span>
-          <div className="d-flex flex-column justify-content-center text-white">
-            <h2 className="text-center">Reserve an Arena</h2>
-          </div>
-          <div className="d-flex">
-            <select className="form-select me-2 rounded-pill" onChange={(e) => setItemId(e.target.value)} value={itemId}>
-              {items.map((item) => (
-                <option key={item.id} value={item.id}>
-                  {item.name}
-                </option>
-              ))}
-            </select>
-          </div>
-          <div className="d-flex justify-content-center w-100 mt-3">
-            <DatePicker selected={startDate} onChange={(date) => setStartDate(date)} />
-          </div>
-          <div className="d-flex justify-content-center w-100 mt-3">
-            <DatePicker selected={endDate} onChange={(date) => setEndDate(date)} />
-          </div>
-          <Container>
-            <Button onClick={openModal}>modal</Button>
-            <Modal showModal={showModal} setShowModal={setShowModal} />
-            <GlobalStyle />
-          </Container>
 
-        </section>
-      </form>
+      <Container>
+        <Button onClick={openModal}>Check here for beautiful Arenas</Button>
+        <Modal showModal={showModal} setShowModal={setShowModal} />
+        <GlobalStyle />
+      </Container>
 
     </div>
   );
